Added PostEmitMetadataArtifactRepresentation
